# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
security:
  - bearerAuth: []
paths:
...
  /wallet/addresses/{id}:
    delete:
      operationId: removeWalletAddress
      tags: [wallet]
      summary: Remove a saved address
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses: { "200": { description: Removed } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Token returned by POST /auth/verify-otp.
...